Change mandated minimum file size to recommended
My understanding is that the max frequency of one file per half hour is enforced by the software and that files that are uploaded more frequently will be dropped. However, I believe that the minimum file size requirement is not enforced. And it would be strange to have a minimum enforced requirement—what if the total amount of customer data is less than 10 MB? Are we going to tell our customers that they can't use customer attributes in that case or that they need to pad the file out with spaces to get it to the limit or something silly like that?